Create context to share logged user
- Added Auth context in the application to manager logged user
  and share methods and props about the logged user

- Included Auth component in the index file to load application
  involved in Auth and share the context

- Simplified the login page, putting the authentication logics
  in the Auth context
