+++ orders-system/src/App.js
@@ -1,16 +1,24 @@
-import React from 'react';
+import React, { useContext } from 'react';
 import { CssBaseline } from '@material-ui/core';
 
 import Pages from 'routes/index';
+import { AuthContext } from 'stories/Auth';
 
 /**
- * Base da aplicação
+ * Base component application
  */
-const App = () => (
-  <>
-    <CssBaseline />
-    <Pages />
-  </>
-);
+const App = () => {
+  const { isLogged, user } = useContext(AuthContext);
+  console.log('isLogged: ', isLogged);
+  console.log('user: ', user);
+
+
+  return (
+    <>
+      <CssBaseline />
+      <Pages />
+    </>
+  );
+};
 
 export default App;
