+++ orders-system/src/index.js
@@ -4,10 +4,20 @@ import ReactDOM from 'react-dom';
 import * as serviceWorker from './serviceWorker';
 
 import App from './App';
+import Auth from './stories/Auth';
 
+// The application is rendered here, the application
+// needs to retrieve authenticated user, so here the
+// component app is involved in the Auth that contains
+// the context about the logged user
+
+// This way the logged user will be accessible all
+// component inside app Auth component
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+    <Auth>
+      <App />
+    </Auth>
   </React.StrictMode>,
   document.getElementById('root'),
 );
