+++ orders-system/src/pages/Login/index.js
@@ -1,7 +1,7 @@
-import React, { useCallback } from 'react';
-import * as Firebase from 'services/Firebase';
-
+import React, { useContext } from 'react';
+import { AuthContext } from 'stories/Auth';
 import * as S from './style';
+
 
 /**
  * Login Page
@@ -12,25 +12,7 @@ import * as S from './style';
  * specific attributes
  */
 const Login = () => {
-  const [userInfo, setUserInfo] = Firebase.useFirebase(null);
-  const { isLogged, user } = userInfo;
-
-  /**
-   * Login method applyng scope Login page
-   */
-  const login = useCallback(() => {
-    Firebase.login();
-  }, []);
-
-  /**
-   * Logout method applyng scope Login page
-   */
-  const logout = useCallback(() => {
-    Firebase.logout()
-      .then(() => {
-        setUserInfo(Firebase.formatUserInfo());
-      });
-  }, [setUserInfo]);
+  const { login } = useContext(AuthContext);
 
   return (
     <S.PageContainer>
@@ -39,26 +21,9 @@ const Login = () => {
           <S.Logo />
         </S.LogoGrid>
         <S.AuthButtonsGrid>
-          { isLogged && (
-            <>
-              <img src={user.picture} width="200px" alt={user.email} />
-              <ul>
-                <li><strong>Uid:</strong> { user.uid} </li>
-                <li><strong>Name:</strong> { user.name} </li>
-                <li><strong>Email:</strong> { user.email} </li>
-                <li><strong>Phone:</strong> { user.phone} </li>
-                <li><strong>Provider:</strong> { user.provider} </li>
-              </ul>
-              <S.Button onClick={logout}>
-                Sair
-              </S.Button>
-            </>
-          )}
-          { !isLogged && (
-            <S.Button onClick={login}>
-              Entrar com GitHub
-            </S.Button>
-          )}
+          <S.Button onClick={login}>
+            Entrar com GitHub
+          </S.Button>
         </S.AuthButtonsGrid>
       </S.PageGrid>
     </S.PageContainer>
